% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_webadmit.R
\name{tidy_webadmit}
\alias{tidy_webadmit}
\title{A function that returns a dataframe with one row per application for a single variable}
\usage{
tidy_webadmit(c.des_var, dfWebadmit, n.variables, dfDesignation, IDvar)
}
\arguments{
\item{c.des_var}{name of the variable to extract}

\item{dfWebadmit}{dataframe extracted from WebAdMIT}

\item{n.variables}{number of designations}

\item{dfDesignation}{dataframe with three columns: every column name in dfWebadmit, an indicator}
}
\description{
A function that returns a dataframe with one row per application for a single variable
}
