% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webadmit_wrapper.R
\name{webadmit_wrapper}
\alias{webadmit_wrapper}
\title{A function to reformat a webadmit export to have one row per application.}
\usage{
webadmit_wrapper(dfWebadmit, IDvar)
}
\arguments{
\item{dfWebadmit}{a dataframe exported from WebAdMIT}

\item{IDvar}{the name of primary identifier variable}
}
\value{
a dataframe with one row per application (designation). Note, all variables will be returned as character type.
}
\description{
A function to reformat a webadmit export to have one row per application.
}
